[AIRLFOW-YYY] Move version command to separate file
